realtek: add support for Zyxel XGS1210-12
This adds support for the Zyxel XGS1210-12, a little brother of the
XGS1250-12 which is already supported in OpenWrt. It is also the managed
version of Zyxel XGS1010-12.

Specifications:
 - Realtek RTL9302B Switch SoC
 - Macronix MX25L12833F (16MB flash)
 - Nanja NT5CC64M16GP-1 (128MB DDR3 SDRAM)
 - RTL8231 GPIO extender to control the port LEDs
 - RTL8218D 8x Gigabit PHY
 - 2x RTL8226 2.5 Gigabit PHY
 - 2x SFP+ 10 Gigabit slot
 - reset button

Power is supplied via a 12V 2A standard barrel connector. At the
right side behind the grid is UART serial connector. A Serial
header can be connected to from the outside of the switch trough
the airvents with a standard 2.54mm header.
Pins are from top to bottom Vcc(3.3V), TX, RX and GND. Serial
connection is via 115200 baud, 8N1.

A reset button is accessble through a hole in the front panel.

Differences to XGS1250-12:
- two SFP+ slots instead of one
- only 2 Multi-Gig Ethernet ports, supporting up to 2500BaseT instead of
  10GBaseT
- passively cooled, no fan

Easy Installation
------------
- Navigate to 'Management' in the OEM web interface and click on
  'Firmware upgrade'.
- Upload the OpenWrt initramfs image and wait till the switch reboots
- Connect to the device through serial and change the U-boot boot command.
  # fw_setenv bootcmd 'rtk network on; boota'
- Reboot, scp the sysupgrade image to /tmp, verify the checksum and flash it:
  # sysupgrade /tmp/openwrt-realtek-rtl930x-zyxel_xgs1210-12-squashfs-sysupgrade.bin
- Upon reboot, you have a functional OpenWrt installation. Leave the bootcmd
  value as is - without 'rtk network on' the switch will fail to initialise

Boot from RAM only
------------------

In contrast to XGS1250-12, U-boot shell is not deactivated by default.
Thus, getting access by hijacking the bootcmd is not necessary.

- Configure TFTP server to provide OpenWrt initramfs image
- Connect to device using serial
- Enter U-boot using Esc when prompted
- Run the following commands (adjust as necessary):
  # setenv loadaddr 0x84f00000
  # rtk network on
  # tftpboot openwrt-realtek-rtl930x-zyxel-xgs1210-12-initramfs-kernel.bin
  # bootm

From RAM-booted OpenWrt, you can also install OpenWrt without using OEM
interface:
  - Run command to persistently change boot command:
    # fw_setenv bootcmd 'rtk network on; boota'
  - Upload sysupgrade image
  - Flash with sysupgrade <sysupgrade.bin>

Web recovery
------------
The XGS1210-12 has a handy web recovery that will load when U-boot does
not find a bootable kernel. In case you would like to trigger the web
recovery manually, partially overwrite the firmware partition with some
zeroes:

  # dd if=/dev/zero of=/dev/mtd5 bs=1M count=2

If you have serial connected you'll see U-boot will start the web recovery
and print it's listening on 192.168.1.1, but by default it seems to be on
the OEM default IP for the switch - 192.168.1.3. The web recovery only
listens on HTTP (80) and *not* on 443 (HTTPS) unlike the web UI.

Return to stock
---------------
You can flash the ZyXEL firmware images to return to stock:

  # sysupgrade -F -n XGS1210-12_Firmware_<VERSION>.bix

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>