x86_64: select CONFIG_PCI_MMCONFIG to fix boot
Add CONFIG_PCI_MMCONFIG to the default kernel config for x86-64 to fix
an issue with Zimaboard. Without this option, the device either crashes
at boot or doesn't even show an output/start to boot properly.

There's no indication so far why this is necessary, however, it is
likely a hardware issue of the Intel N3450.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>