realtek: add support for Netgear MS510TXM
The Netgear MS510TXM is a professional managed switch with the following
specs:

* SoC: Realtek RTL9313 Switch SoC
* RAM: Winbond W623GU6NB-11 (256MB DDR3L RAM)
* Flash:
  * MX25L25645G (32MB)
  * MX25U4033E (for BCM84888 firmware)
* Network Ports:
  * 4x 2.5GBase-T (4x RTL8221B)
  * 4x 10GBase-T (1x Broadcom BCM84888)
  * 2x SFP+
* Serial: 3V3 logic, 115200 8N1, populated pins on PCB
* LEDs:
  * Dual-Color LED per network port
  * Power-LED
  * Fan-LED
* Buttons:
  * Reset-Button
* Cooling: 2x 4-pin controllable fans
* Power: C13 socket (integrated power supply)
* Monitoring:
  * TMP1075 (I2C temperature sensor)
  * ADT747 (I2C fan controller)
* Bootloader: U-Boot Realtek SDK with networking

DISCLAIMER: BCM84888 (and thus 10G ports) not yet supported!!!

Case has to be opened to access the serial header. Unscrew the front
panel first (4 screws) and then the cover panel (3 screws).

Installation
------------

1. Download/build the initramfs image for the device.
2. Connect to the switch via network cable and open the web interface of
   the vendor firmware.
3. Go to firmware upgrade and upload the OpenWrt initramfs image (vendor
   firmware upgrade is a simple uimage file too.)
4. Wait till the switch reboots, it should now boot into OpenWrt.
5. Connect to the device via serial.
6. Adjust boot command with:
  > fw_setenv bootcmd 'rtk network on; boota'
7. Reboot, scp the sysupgrade image and do a sysupgrade.
8. Device should now boot OpenWrt with functioning networking.

Initramfs boot
--------------

In theory, the device can be ram-booted. However, this is similar to the
Zyxel XGS1250-12. Access to U-Boot console is restricted in such a way
that you cannot interrupt autoboot. You need to "flash" the initramfs
first, from there set a bogus bootcmd to force U-Boot to drop into
a shell and can then use 'tftpboot' as usual. After that you can
reinstall the OEM firmware. But as long as the bootcmd is bogus, you
need to manually boot the device via serial.

Revert to stock
---------------

TBD

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>