ramips: add erofs to bootargs to fix boot with EROFS rootfs
Add erofs as another supported type to the 'rootfstype' boot argument
that is used for several ramips subtargets and devices.

While images with EROFS as rootfs can be built without an issue,
running them is going to fail because 'rootfstype' is limited to
squashfs and jffs2, not considering the erofs driver.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>