realtek: backport gpio-line-mux driver
Backport the upstream 'gpio-line-mux' driver which allows to provide a
1-to-many mapping between one physical GPIO and multiple virtual GPIOs,
based on a multiplexer.

Although there hasn't been any need for this so far, with the Zyxel
XS1930-10 there's at least one device which requires this. The SFP
signals 'RX_LOS', 'MOD_ABS' und 'TX_FAULT' do not have dedicated GPIOs
each but all use a single GPIO which is multiplexed. Depending on the
multiplexer state the GPIO line is connected to one of the signals.

Since the SFP driver needs single GPIOs for the single signals, this
adapter drivers fills the gap to make both work together.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>