realtek: add gpio-line-mux driver
Add a new driver which allows to split a physical GPIO into multiple
virtual GPIOs by using a multiplexer.

Although there hasn't been any need for this so far, with the Zyxel
XS1930-10 there's at least one device which requires this. The SFP
signals 'RX_LOS', 'MOD_ABS' und 'TX_FAULT' do not have dedicated GPIOs
each but all use a single GPIO which is multiplexed. Depending on the
multiplexer state the GPIO line is connected to one of the signals.

Since the SFP driver needs single GPIOs for the single signals, this
adapter drivers fills the gap to make both work together.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>