realtek: pcs: rtl931x: move rtpcs_931x_sds_rx_reset up
Move the function 'rtpcs_931x_sds_rx_reset' up in the code to have it
within a logical block which deals with SerDes powering, reset, mode
setting. Up to now, it was somewhat homeless below the CMU stuff.

This has no functional changes and is just for the purpose of having
some structure in the driver code.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>