realtek: pcs: rtl931x: drop top-level XGMII handling
Drop the XGMII handling in the RTL931X SerDes setup. XGMII is parallel
mode and not supported by Realtek SerDes. The mode was used incorrectly
as an alias for the proprietary XSGMII mode. However, the XSGMII mode is
apparently not in a working state. Thus, drop this confusion handling
for now. A proper implementation of XSGMII mode should be provided
later.

Mark the function rtpcs_931x_sds_mii_mode_set as unused because XGMII
was the only user of this and the call is dropped here.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>