realtek: pcs: rtl931x: drop unneeded and bad helper
Drop the unneeded and bad helper rtpcs_931x_sds_disable. It is bad
because it name implies that it disable a SerDes completely. However, it
only sets the MII mode to off. Dropping this and using a direct call of
rtpcs_931x_sds_mii_set_mode with RTPCS_SDS_MODE_OFF is much better for
clarity. As the helper is only used once, it isn't actually needed.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>