realtek: pcs: rtl931x: fix clearing symbol errors
Fix the implementation to clear symbol errors. Up to now, it was
basically incomplete and rather non-functional because important modes
weren't handled. For this, add another function to read symbol errors as
this is needed to clear them for some modes (i.e. clear-on-read).

This is based on the corresponding SDK function [1].

While at it, slightly rename the function to adhere to a better naming
scheme.

[1] https://github.com/plappermaul/realtek-doc/blob/69d2890a2e2d7a03df6e40e2cd2c32ff5b074dcf/sources/rtk-dms1250/src/hal/phy/phy_rtl9310.c#L703

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>