realtek: pcs: rtl931x: fix SerDes mode setting
The SerDes mode setting at the end of rtpcs_931x_setup_serdes is
currently broken although it is mostly similar to what the SDK does [1].
It prevents several modes from being set, especially fiber modes. Thus,
this is one (if not the only issue) for currently missing SFP support.

Add a small helper function which takes cares of setting the mode, to
keep the probably still valid different procedure when using XSGMII
mode. Only this helper is called in the big rtpcs_931x_setup_serdes to
keep it clean there.

As a functional change, call fiber mode setting in every case, not just
for SGMII, QSGMII and USXGMII. We can assume the SDK is misleading in
this case, either accidentially or on purpose. This makes SFP modules
at least work in principle with link detection. However, there still
seem to be device-specific issues which need to be fixed later.

[1] https://github.com/plappermaul/realtek-doc/blob/f7f85ffc14c6e575ad94f2b427106de533b888fe/sources/rtk-dms1250/src/dal/mango/dal_mango_construct.c#L2266

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>