realtek: pcs: rtl931x: add port media handling
SFP modules still do not work that well across different devices. One
missing piece seems to be the bunch of magic values and bits set by the
SDK depending on which media is used on a port.

Take over code from the SDK for port media handling [1]. This applies
different sequences depending on whether it's 10G fiber, 1G fiber or DAC
cables to make it work best for each variant. Place the call to that
code below the configuration of a SerDes mode but before actually
activating that mode and powering on the SerDes. The SDK and our code
for RTL930x do that similarly.

Though we do not have any notion of media in pcs_config right now, do
similar to some SDK versions and assume 10G fiber for most variants and
1G fiber for 1000Base-X and SGMII.

[1] https://github.com/plappermaul/realtek-doc/blob/51c3390e0e47ac068fd0100e7fa6c10355042d16/sources/rtk-dms1250/src/hal/phy/phy_rtl9310.c#L2302

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>